# openapi.json
{"openapi":"3.1.0","info":{"title":"Urja Meter Wrapper API","description":"A production-grade REST API wrapper over the legacy Urja Meter Ops portal.","version":"1.0.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Simulates user login on the API wrapper.\nVerifies connection credentials to the underlying legacy portal.","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters":{"get":{"tags":["Meters"],"summary":"List Meters","description":"Retrieve normalized list of all energy meters parsed from legacy portal.","operationId":"list_meters_api_v1_meters_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MeterBase"},"type":"array","title":"Response List Meters Api V1 Meters Get"}}}}}}},"/api/v1/meters/{meter_id}":{"get":{"tags":["Meters"],"summary":"Get Meter Detail","description":"Retrieve detailed specs for a specific meter by its ID, merging hierarchy details.","operationId":"get_meter_detail_api_v1_meters__meter_id__get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeterDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}/consumption":{"get":{"tags":["Meters"],"summary":"Get Meter Consumption","description":"Retrieve historical active energy, reactive energy, and max demand readings.","operationId":"get_meter_consumption_api_v1_meters__meter_id__consumption_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsumptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/hierarchy":{"get":{"tags":["Hierarchy"],"summary":"Get Hierarchy","description":"Retrieve normalized grid layout network tree (Zones, Circles, Transformers, Meters).","operationId":"get_hierarchy_api_v1_hierarchy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/HierarchyNode"},"type":"array","title":"Response Get Hierarchy Api V1 Hierarchy Get"}}}}}}},"/health":{"get":{"tags":["System"],"summary":"Health Check","description":"System health check including connectivity checks to Redis and Legacy portal.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ConsumptionReading":{"properties":{"date":{"type":"string","title":"Date"},"energy_kwh":{"type":"number","title":"Energy Kwh"},"reactive_kvarh":{"type":"number","title":"Reactive Kvarh"},"demand_kw":{"type":"number","title":"Demand Kw"}},"type":"object","required":["date","energy_kwh","reactive_kvarh","demand_kw"],"title":"ConsumptionReading"},"ConsumptionResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"serial_number":{"type":"string","title":"Serial Number"},"readings":{"items":{"$ref":"#/components/schemas/ConsumptionReading"},"type":"array","title":"Readings"}},"type":"object","required":["meter_id","serial_number","readings"],"title":"ConsumptionResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HierarchyNode":{"properties":{"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type"},"children":{"anyOf":[{"items":{"$ref":"#/components/schemas/HierarchyNode"},"type":"array"},{"type":"null"}],"title":"Children"},"meter":{"anyOf":[{"$ref":"#/components/schemas/MeterNode"},{"type":"null"}]}},"type":"object","required":["name","type"],"title":"HierarchyNode"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"token":{"type":"string","title":"Token"}},"type":"object","required":["success","message","token"],"title":"LoginResponse"},"MeterBase":{"properties":{"id":{"type":"string","title":"Id"},"serial_number":{"type":"string","title":"Serial Number"},"location":{"type":"string","title":"Location"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"phase":{"type":"string","title":"Phase"},"installation_date":{"type":"string","title":"Installation Date"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","serial_number","location","latitude","longitude","phase","installation_date","status"],"title":"MeterBase"},"MeterDetail":{"properties":{"id":{"type":"string","title":"Id"},"serial_number":{"type":"string","title":"Serial Number"},"location":{"type":"string","title":"Location"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"phase":{"type":"string","title":"Phase"},"installation_date":{"type":"string","title":"Installation Date"},"status":{"type":"string","title":"Status"},"voltage":{"type":"number","title":"Voltage"},"current":{"type":"number","title":"Current"},"power_factor":{"type":"number","title":"Power Factor"},"zone":{"type":"string","title":"Zone"},"circle":{"type":"string","title":"Circle"},"substation":{"type":"string","title":"Substation"},"transformer":{"type":"string","title":"Transformer"}},"type":"object","required":["id","serial_number","location","latitude","longitude","phase","installation_date","status","voltage","current","power_factor","zone","circle","substation","transformer"],"title":"MeterDetail"},"MeterNode":{"properties":{"id":{"type":"string","title":"Id"},"serial_number":{"type":"string","title":"Serial Number"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","serial_number","status"],"title":"MeterNode"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}